First x86_64 commit - Hello, Wolrd!; Working on tty

+++ kernel/mm/pmm.c
@@ -7,21 +7,21 @@
 #define BITMAP_SIZE 131072
 
 static uint8_t pmm_bitmap[BITMAP_SIZE];
-static uint32_t max_blocks = 0;
+static uintptr_t max_blocks = 0;
 
-extern uint32_t _end;
+extern uintptr_t _end;
 
-static inline void bitmap_set(uint32_t bit)
+static inline void bitmap_set(uintptr_t bit)
 {
   pmm_bitmap[bit / 8] |= (1 << (bit % 8));
 }
 
-static inline void bitmap_clear(uint32_t bit)
+static inline void bitmap_clear(uintptr_t bit)
 {
   pmm_bitmap[bit / 8] &= ~(1 << (bit % 8));
 }
 
-static inline int bitmap_test(uint32_t bit)
+static inline int bitmap_test(uintptr_t bit)
 {
   return pmm_bitmap[bit / 8] & (1 << (bit % 8));
 }
@@ -36,33 +36,33 @@ void pmm_init(multiboot_info_t* mbd)
     return;
   }
 
-  multiboot_memory_map_t* mmap = (multiboot_memory_map_t*)mbd->mmap_addr;
-  uint32_t mmap_end = mbd->mmap_addr + mbd->mmap_length;
+  multiboot_memory_map_t* mmap = (multiboot_memory_map_t*)(uintptr_t)mbd->mmap_addr;
+  uintptr_t mmap_end = mbd->mmap_addr + mbd->mmap_length;
 
-  while ((uint32_t)mmap < mmap_end)
+  while ((uintptr_t)mmap < mmap_end)
   {
     if (mmap->type == 1) // 1 => memory available
     {
-      uint32_t start_addr = mmap->base_addr_low;
-      uint32_t length = mmap->length_low;
+      uintptr_t start_addr = mmap->base_addr_low;
+      uintptr_t length = mmap->length_low;
 
-      uint32_t start_block = start_addr / PMM_BLOCK_SIZE;
-      uint32_t blocks = length / PMM_BLOCK_SIZE;
+      uintptr_t start_block = start_addr / PMM_BLOCK_SIZE;
+      uintptr_t blocks = length / PMM_BLOCK_SIZE;
 
-      for (uint32_t i = 0; i < blocks; i++)
+      for (uintptr_t i = 0; i < blocks; i++)
         bitmap_clear(start_block + i);
 
-      uint32_t end_block = start_block + blocks;
+      uintptr_t end_block = start_block + blocks;
       if (end_block > max_blocks)
         max_blocks = end_block;
     }
-    mmap = (multiboot_memory_map_t *) ((uint32_t)mmap + mmap->size + sizeof(mmap->size));
+    mmap = (multiboot_memory_map_t *) ((uintptr_t)mmap + mmap->size + sizeof(mmap->size));
   }
 
-  uint32_t kernel_end_addr = (uint32_t)&_end;
-  uint32_t kernel_blocks = (kernel_end_addr / PMM_BLOCK_SIZE) + 1;
+  uintptr_t kernel_end_addr = (uintptr_t)&_end;
+  uintptr_t kernel_blocks = (kernel_end_addr / PMM_BLOCK_SIZE) + 1;
 
-  for (uint32_t i = 0; i < kernel_blocks; i++)
+  for (uintptr_t i = 0; i < kernel_blocks; i++)
     bitmap_set(i);
 
   printf("[ INFO ] Mist: PMM initialized. Max blocks: %d (4KB)\n", max_blocks);
@@ -70,18 +70,18 @@ void pmm_init(multiboot_info_t* mbd)
 
 void* pmm_alloc_block(void)
 {
-  for (uint32_t i = 0; i < max_blocks; i++)
+  for (uintptr_t i = 0; i < max_blocks; i++)
     if (!bitmap_test(i))
     {
       bitmap_set(i);
-      return (void*)(i * PMM_BLOCK_SIZE);
+      return (void*)(uintptr_t)(i * PMM_BLOCK_SIZE);
     }
   return NULL;
 }
 
 void pmm_free_block(void* ptr)
 {
-  uint32_t addr = (uint32_t)ptr;
-  uint32_t block = addr / PMM_BLOCK_SIZE;
+  uintptr_t addr = (uintptr_t)ptr;
+  uintptr_t block = addr / PMM_BLOCK_SIZE;
   bitmap_clear(block);
 }
